pyproject.toml
[tool.poetry]
name = "portray"
version = "0.0.4"
description = "Your Project with Great Documentation"
authors = ["Timothy Crosley <timothy.crosley@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pdoc3 = "^0.6.3"
hug = "^2.5"
mkdocs = "^1.0"
toml = "^0.10.0"
mkdocs-material = "^4.4"
GitPython = "^3.0"
pymdown-extensions = "^6.0"

[tool.poetry.dev-dependencies]
vulture = "^1.0"
bandit = "^1.6"
pytest = "^5.1"
safety = "^1.8"
isort = "^4.3"
flake8-bugbear = "^19.8"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
mypy = "^0.720.0"
ipython = "^7.7"
pytest-cov = "^2.7"

[tool.poetry.scripts]
portray = "portray.cli:__hug__.cli"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

